requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
python-dotenv==1.0.1
psycopg[binary]==3.2.1
pydantic==2.9.2
requests==2.32.3
numpy==1.26.4
pandas==2.2.2
sentence-transformers==3.0.1
pypdf==5.0.1
